requirements.txt
aniso8601==4.0.1
Click==7.0
Flask==1.0.2
Flask-RESTful==0.3.7
itsdangerous==1.1.0
Jinja2==2.10
joblib==0.13.0
MarkupSafe==1.1.0
numpy==1.15.4
pandas==0.23.4
python-dateutil==2.7.5
pytz==2018.7
scikit-surprise==1.0.6
scipy==1.2.0
six==1.12.0
Werkzeug==0.14.1
